requirements.txt
requests
python-decouple
quixstreams
psycopg2-binary
python-dotenv